Made more functions static in OpenFFA, removed some dead code. PMenu code no longer uses dynamically allocated memory. Fixed player ID display for chasecam spectators. Player ID display code now works currectly underwater even when server has ‘map_allsolid_bug’ variable set to 0. Implemented wave flood protection, added ‘flood_waves’ variable and friends. Implemented ‘settings’, ‘highscores’ and ‘commands’ client commands. Moved voting code into g_vote.c. Attempt to fix ‘shooting through bodies’ bug by the means of game itself. Made global sound messages reliable. Send MZ_LOGIN effect on the newly connected player, but only to this player.

+++ g_utils.c
@@ -553,7 +553,7 @@ qboolean G_KillBox (edict_t *ent)
     if( level.intermission_framenum ) {
         return qfalse;
     }
-    if (g_bugs->value < 2)
+    if ((int)g_bugs->value < 2)
     {
         edict_t *touch[MAX_EDICTS];
         int     count;
